requirements.txt
feedparser
python-dotenv
chromadb
sentence_transformers
trafilatura